# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AccountCurrency = typeof AccountCurrency[keyof typeof AccountCurrency];


export const AccountCurrency = {
  USD: 'USD',
} as const;

export interface Account {
  id: string;
  balanceCents: number;
  currency: AccountCurrency;
}

export interface ApiKey {
  id: string;
  name: string;
  prefix: string;
  createdAt: string;
}

export interface KeyInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
}

export interface CreatedKey {
  key: ApiKey;
  secret: string;
}

export interface ModelPrice {
  id: string;
  provider: string;
  model: string;
  inputUsdPerMillion: string;
  outputUsdPerMillion: string;
}

export interface RequestRecord {
  id: string;
  createdAt: string;
  keyName: string;
  model: string;
  transport: string;
  status: string;
  inputTokens: number;
  outputTokens: number;
  chargedCents: number;
  /** @nullable */
  error?: string | null;
}

export interface Workspace {
  account: Account;
  keys: ApiKey[];
  requests: RequestRecord[];
  requestCount: number;
  spentCents: number;
}