# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountCurrency } from './accountCurrency';

export interface Account {
  id: string;
  balanceCents: number;
  currency: AccountCurrency;
}