# lib/api-zod/src/generated/types/createdKey.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKey } from './apiKey';

export interface CreatedKey {
  key: ApiKey;
  secret: string;
}