# lib/api-zod/src/generated/types/keyInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KeyInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
}