# lib/api-zod/src/generated/types/modelPrice.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ModelPrice {
  id: string;
  provider: string;
  model: string;
  inputUsdPerMillion: string;
  outputUsdPerMillion: string;
}