# lib/api-zod/src/generated/types/requestRecord.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RequestRecord {
  id: string;
  createdAt: Date;
  keyName: string;
  model: string;
  transport: string;
  status: string;
  inputTokens: number;
  outputTokens: number;
  chargedCents: number;
  /** @nullable */
  error?: string | null;
}