# lib/api-zod/src/generated/types/workspace.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Account } from './account';
import type { ApiKey } from './apiKey';
import type { RequestRecord } from './requestRecord';

export interface Workspace {
  account: Account;
  keys: ApiKey[];
  requests: RequestRecord[];
  requestCount: number;
  spentCents: number;
}